+++ packages/extensions/lib/src/ai/speech_to_text/configure_options_speech_to_text_client.dart
@@ -1,5 +1,6 @@
 import '../../system/threading/cancellation_token.dart';
 import 'delegating_speech_to_text_client.dart';
+import 'speech_to_text_response_update.dart';
 import 'speech_to_text_client.dart';
 
 /// A delegating speech-to-text client that applies configuration to
@@ -28,7 +29,7 @@ class ConfigureOptionsSpeechToTextClient extends DelegatingSpeechToTextClient {
   );
 
   @override
-  Stream<SpeechToTextResponse> getStreamingText({
+  Stream<SpeechToTextResponseUpdate> getStreamingText({
     required Stream<List<int>> stream,
     SpeechToTextOptions? options,
     CancellationToken? cancellationToken,