+++ packages/extensions/lib/src/ai/speech_to_text/delegating_speech_to_text_client.dart
@@ -1,5 +1,6 @@
 import '../../system/threading/cancellation_token.dart';
 import 'speech_to_text_client.dart';
+import 'speech_to_text_response_update.dart';
 
 /// A [SpeechToTextClient] that delegates all calls to an inner client.
 ///
@@ -26,7 +27,7 @@ abstract class DelegatingSpeechToTextClient implements SpeechToTextClient {
   );
 
   @override
-  Stream<SpeechToTextResponse> getStreamingText({
+  Stream<SpeechToTextResponseUpdate> getStreamingText({
     required Stream<List<int>> stream,
     SpeechToTextOptions? options,
     CancellationToken? cancellationToken,